requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Werkzeug==2.3.0
Jinja2==3.1.2
MarkupSafe==2.1.2
click==8.1.3
itsdangerous==2.1.2